feat: add article search service layer
Building up to the article search endpoint

Issues:
closes #136
